# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "FinZ LMS API",
        "description": "API documentation for FinZ LMS backend",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/v1/admin/disbursals/pending": {
            "get": {
                "tags": [
                    "DisbursalSettlement"
                ],
                "summary": "pendingDisbursals DisbursalSettlement",
                "operationId": "cd9a437aebc0f7556d2c6a9c58a7cadd",
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/admin/disbursals/trigger-batch": {
            "post": {
                "tags": [
                    "DisbursalSettlement"
                ],
                "summary": "triggerBatchDisbursal DisbursalSettlement",
                "operationId": "2ded653309bb59efb944d8c5d20e899a",
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/admin/settlements/batches": {
            "get": {
                "tags": [
                    "DisbursalSettlement"
                ],
                "summary": "settlementBatches DisbursalSettlement",
                "operationId": "8bab8aaeb9d81b3f5fd38607e38eec4a",
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/admin/settlements/batches/{batch_id}/entries": {
            "get": {
                "tags": [
                    "DisbursalSettlement"
                ],
                "summary": "settlementEntries DisbursalSettlement",
                "operationId": "3f0b18c0a6d75a12f75763757fd3278e",
                "parameters": [
                    {
                        "name": "batch_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/admin/settlements/batches/{batch_id}/download": {
            "get": {
                "tags": [
                    "DisbursalSettlement"
                ],
                "summary": "downloadSettlement DisbursalSettlement",
                "operationId": "f338e0f611a8747a59de0833ca31465f",
                "parameters": [
                    {
                        "name": "batch_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/admin/settlements/entries/{entry_id}/dispute": {
            "post": {
                "tags": [
                    "DisbursalSettlement"
                ],
                "summary": "disputeSettlement DisbursalSettlement",
                "operationId": "30d7c9d086262ab1ddf0613b7d245886",
                "parameters": [
                    {
                        "name": "entry_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/admin/fraud-alerts": {
            "get": {
                "tags": [
                    "FraudAlert"
                ],
                "summary": "index FraudAlert",
                "operationId": "1c101bfe14cfe4edb6dedaf594f8a9bb",
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/admin/fraud-alerts/{id}/block": {
            "post": {
                "tags": [
                    "FraudAlert"
                ],
                "summary": "block FraudAlert",
                "operationId": "e9a60b367893f342406b5e58c0b9d7db",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/admin/fraud-alerts/{id}/unblock": {
            "post": {
                "tags": [
                    "FraudAlert"
                ],
                "summary": "unblock FraudAlert",
                "operationId": "660c1651cc9b671eebf355bb7072e474",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "requestBody": {
                    "description": "Optional. ID of another existing category to reassign all products to before archiving. If provided, the ID must exist in the categories table.",
                    "required": false,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "reassign_to": {
                                        "description": "Must be a valid existing Category ID",
                                        "type": "integer",
                                        "example": null
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/admin/fraud-alerts/{id}/escalate": {
            "post": {
                "tags": [
                    "FraudAlert"
                ],
                "summary": "escalate FraudAlert",
                "operationId": "49f932969c5c7f52cc792f90d6b99dac",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/admin/fraud-alerts/stats/heatmap": {
            "get": {
                "tags": [
                    "FraudAlert"
                ],
                "summary": "heatmap FraudAlert",
                "operationId": "808744af97e6e403889a17dba026052a",
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
...
}